requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
